+++ lib/eventual-single-builder.dart
@@ -2,13 +2,13 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:eventual/eventual-notifier.dart';
 
-/// EventualBuilders provides up to four different builders that are used according to the
+/// EventualSingleBuilder provides up to four different builders that are used according to the
 /// current state of `notifier`.
 /// - If `isLoading` evaluates to `true`, then `loadingBuilder` will be built (if present).
 /// - If `hasError` evaluates to `true`, then `errorBuilder` will be built (if present).
 /// - If `hasValue` evaluates to `false`, then `emptyBuilder` will be built (if present).
 /// - In any other case, `builder` will be built.
-class EventualBuilders extends StatefulWidget {
+class EventualSingleBuilder extends StatefulWidget {
   final EventualNotifier notifier;
   final Widget Function(BuildContext, EventualNotifier, Widget) loadingBuilder;
   final Widget Function(BuildContext, EventualNotifier, Widget) errorBuilder;
@@ -16,7 +16,7 @@ class EventualBuilders extends StatefulWidget {
   final Widget Function(BuildContext, EventualNotifier, Widget) builder;
   final Widget child;
 
-  EventualBuilders(
+  EventualSingleBuilder(
       {@required this.notifier,
       this.loadingBuilder,
       this.errorBuilder,
@@ -27,10 +27,10 @@ class EventualBuilders extends StatefulWidget {
       : super(key: key);
 
   @override
-  _EventualBuildersState createState() => _EventualBuildersState();
+  _EventualSingleBuilderState createState() => _EventualSingleBuilderState();
 }
 
-class _EventualBuildersState extends State<EventualBuilders> {
+class _EventualSingleBuilderState extends State<EventualSingleBuilder> {
   int _buildCount = 0;
 
   @override
@@ -54,7 +54,7 @@ class _EventualBuildersState extends State<EventualBuilders> {
   }
 
   @override
-  void didUpdateWidget(EventualBuilders oldWidget) {
+  void didUpdateWidget(EventualSingleBuilder oldWidget) {
     if (oldWidget.notifier != widget.notifier) {
       oldWidget.notifier.removeListener(_changeHandler);