+++ lib/eventual.dart
@@ -3,4 +3,4 @@ library eventual;
 export "package:eventual/eventual-value.dart";
 export "package:eventual/eventual-notifier.dart";
 export "package:eventual/eventual-builder.dart";
-export "package:eventual/eventual-builders.dart";
+export "package:eventual/eventual-single-builder.dart";